21-Data Transfer Object (DTO) Introduced to USIM's Constructor
I am following step by step refactoring process.
In general SIM the MNO profile gets inserted during the UICC Manufacturing Process. That is why I and Inserting the MNO profile directly in the USIM constructor for now. Soon there will be a Single Responsibility Contract for Creating and managing the MNO. So we can dynamically adopt and switch multiple MNO profiles in the eSIM Contract.